refactor analytics options under the plugin type
